Make grid stay black after it turns black
There was a bug in the previous commit. The code there assume RGB of
255 255 255 is black and RGB of 0 0 0 is white. However, this is the
opposite of what is happeing. So the code changes made sure to refect
that.

There was also nother bug regarding the initial background color of the
grid. From the previous commit, the background color is not set to anything
at all but the background color is white. When the JS extracted this bg color
the result is RGB 0 0 0. But for some reason, after the background color
is explicitly defined, the RBG 0 0 0 will replecent black and not white.
Therefore, to remedy this problem, the bg color is explicitly set to RGB
255 255 255 from the moment it is created.